+++ R/format_sumstats.R
@@ -140,6 +140,10 @@
 #' along with effect and z-score columns like Beta? Default TRUE.
 #' @param bi_allelic_filter Binary Should non-biallelic SNPs be removed. Default
 #' is TRUE.
+#' @param flip_frq_as_biallelic Binary Should non-bi-allelic SNPs frequency 
+#' values be flipped as 1-p despite there being other alternative alleles? 
+#' Default is FALSE but if set to TRUE, this allows non-bi-allelic SNPs to be 
+#' kept despite needing flipping.
 #' @param snp_ids_are_rs_ids Binary Should the supplied SNP ID's be assumed to
 #' be RSIDs. If not, imputation using the SNP ID for other columns like
 #' base-pair position or chromosome will not be possible. If set to FALSE, the
@@ -249,6 +253,7 @@ format_sumstats <- function(path,
                             allele_flip_z = TRUE,
                             allele_flip_frq = TRUE,
                             bi_allelic_filter = TRUE,
+                            flip_frq_as_biallelic = FALSE,
                             snp_ids_are_rs_ids = TRUE,
                             remove_multi_rs_snp = FALSE,
                             frq_is_maf = TRUE,
@@ -336,6 +341,7 @@ format_sumstats <- function(path,
             allele_flip_z = allele_flip_z,
             allele_flip_frq = allele_flip_frq,
             bi_allelic_filter = bi_allelic_filter,
+            flip_frq_as_biallelic = flip_frq_as_biallelic,
             snp_ids_are_rs_ids = snp_ids_are_rs_ids,
             remove_multi_rs_snp = remove_multi_rs_snp,
             frq_is_maf = frq_is_maf,
@@ -702,6 +708,7 @@ format_sumstats <- function(path,
                 allele_flip_z = allele_flip_z,
                 allele_flip_frq = allele_flip_frq,
                 bi_allelic_filter = bi_allelic_filter,
+                flip_frq_as_biallelic = flip_frq_as_biallelic,
                 imputation_ind = imputation_ind,
                 log_folder_ind = log_folder_ind,
                 check_save_out = check_save_out,